move logger to log package
